backend/requirements.txt
fastapi>=0.115.0
uvicorn[standard]>=0.34.0
requests>=2.32.0
neo4j>=5.28.0
pillow>=11.1.0
python-multipart>=0.0.20
